pyproject.toml
[tool.poetry]
name = "smartmonkey-highway"
description = "Python library for use the SmartMonkey Highway services"
version = "0.0.1"
authors = ["Urbvan Backend Team <tec@urbvan.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pre-commit = "^2.10.1"
httpretty = "^1.0.5"

[tool.poetry.urls]
repository = "https://github.com/urbvantransit/smartmonkey-highway"
"API documentation" = "https://highwayservices.smartmonkey.io/documentation"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | compose
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
